requirements.txt
pydantic>=2.0
typer[all]>=0.9.0
PyYAML>=6.0
pandas>=2.0
openpyxl>=3.1
matplotlib>=3.7
PySide6>=6.5
pytest>=7.0
